+++ src/libromdata/Console/GameCube.cpp
@@ -626,22 +626,10 @@ const char *GameCubePrivate::wii_getCryptoStatus(WiiPartition *partition)
 		}
 	}
 
-	if (res == KeyManager::VerifyResult::OK) {
+	if (res == KeyManager::VerifyResult::IncrementingValues) {
 		// Debug discs may have incrementing values instead of a
 		// valid update partition.
-		static const uint8_t incr_vals[32] = {
-			0x00,0x00,0x00,0x00, 0x00,0x00,0x00,0x04,
-			0x00,0x00,0x00,0x08, 0x00,0x00,0x00,0x0C,
-			0x00,0x00,0x00,0x10, 0x00,0x00,0x00,0x14,
-			0x00,0x00,0x00,0x18, 0x00,0x00,0x00,0x1C,
-		};
-
-		uint8_t data[32];
-		size_t size = partition->seekAndRead(0, data, sizeof(data));
-		if (size == sizeof(incr_vals) && !memcmp(data, incr_vals, sizeof(data))) {
-			// Found incrementing values.
-			return C_("GameCube", "Incrementing values");
-		}
+		return C_("GameCube", "Incrementing values");
 	}
 
 	const char *err = KeyManager::verifyResultToString(res);